Value class parameters can be private[this].
A discussion on gitter led me to believe that the restriction
was based on ease of implementation rather than any principle
underlying value classes. Therefore, drop that restriction in
2.13 mode and onward.

There are a few problems caused by just removing the error:
  - there's no accessor generated during method synthesis,
    and that's mostly fine if we're willing to return the
    paramaccessor from `derivedValueClassUnbox`, except...
  - the generated `equals` method needs to unbox another
    instance of the same value class, which can't be done
    with the LOCAL paramaccessor, and...
  - adapting the paramaccessor call inserts parens when
    generating the unbox method, which turns into an apply()
    invocation on the non-NullaryMethodTyped paramaccessor

The `other.unbox` invocation in the `equals` method will
pass runtime muster, so by generating an already-typed
selection we can avoid failing the accessibility check,
and by relying on empty-params adaptation we can stop
wrapping the value class unboxing cast in an `Apply`.

--- but why, you ask? ---
One of the biggest use cases for value classes is zero-cost
extension methods. However, the accessor `val` is considered
a method on the value class, so searching in `adaptToMember`
will find the `val`, if it's not made `private`. However, this
makes the syntax for declaring the extension class much more
verbose, and introduces binary compatibility issues as well as
potentially unwanted extension methods.