Remove the NameOps extension methods.
Now that the `String => Name` implicit conversion is gone,
there's no* need for a `NameOps` trait to put the methods
that would have otherwise ambiguated the same-named methods
on various other `String` wrappers. At least that's what I
imagine the motivation to be.

The comment about `NameOps` being a wish-it-were-a-valueclass
led me to just get rid of it entirely. I don't expect a tangible
performance improvement by removing that allocation, but I can't
imagine that this doesn't make things ever so infinitesimally
better.

* Inside `nme.segments`, both the `wrapString` and `createNameType`
  conversions are eligible to give the `drop` method and its friends.
  One very small shadowing `def` inserted.