Use a bytecode switch for matches on Java enums.
Currently, we emit a chained if statement, which can become unwieldy
and slow in the face of large enums. This replaces it with (approximately)
what Java does: compute (once) the mapping from compile-time enum
ordinals to run-time enum ordinals, and switch on those. Java generates
an inner class with a static array, one per switch statement. We get
the benefit of forcing everyone to link to the current version of
scala-library.jar, so use a bootstrap method to make the array instead.

The backend has been updated to deal with varargs bootstrap methods
in a slightly silly way (we always assume that all the args are
static args if so...). I'm not sure what a better story would be
here, and it shouldn't break any intrepid macro writers' code (they
couldn't have used varargs bootstrap methods anyways).

The checkfile changes are because I moved the `RegularSwitchMaker`
construction inside of an `if` block: it allocates one fresh name
for the `default` label in the constructor.

Fixes scala/scala-dev#283.